requirements.txt
matplotlib
pandas
numpy
pynrrd
pydicom
SimpleITK
imageio
natsort
tqdm
plotly
dash